requirements.txt
numpy<2.0, >=1.16.0
scipy
Pillow
cython
matplotlib
scikit-image
tensorflow-gpu>=1.3.0, <2.0
keras==2.1.5
opencv-python
h5py
imgaug
IPython[all]
